feat: added schemas to the same file
